Moved NavigationController to its own file.
